# docs/openapi/v1/openapi.yaml
openapi: 3.0.3
info:
  title: MoveONgs API
  description: |
    Especificação versionada dos principais endpoints REST do backend MoveONgs.
    Use este documento para integrações de clientes, geração de SDKs e revisões
    de conformidade. Cada alteração de contrato deve ser refletida nesta versão.
  version: 1.0.0
servers:
  - url: https://api.dev.moveongs.local
    description: Ambiente de desenvolvimento
  - url: https://api.staging.moveongs.org
    description: Homologação
  - url: https://api.moveongs.org
    description: Produção
...
security:
  - bearerAuth: []
paths:
...
  /attachments:
    get:
      tags: [Attachments]
      summary: Lista anexos vinculados
      parameters:
        - name: ownerType
          in: query
          required: true
          schema:
            type: string
        - name: ownerId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Anexos retornados
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Attachment'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags: [Attachments]
      summary: Faz upload de arquivo
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                ownerType:
                  type: string
                ownerId:
                  type: string
                  format: uuid
      responses:
        '201':
          description: Anexo cadastrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  attachment:
                    $ref: '#/components/schemas/Attachment'
        '400':
          description: Upload inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedError:
      description: Token ausente ou inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        details:
          nullable: true
          description: Informações adicionais sobre o erro
...
    Attachment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ownerType:
          type: string
        ownerId:
          type: string
          format: uuid
        fileName:
          type: string
        mimeType:
          type: string
          nullable: true
        sizeBytes:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
...